proabbly fix

+++ system/systemPlugins/core.py
@@ -150,34 +150,6 @@ except:
 				print("Fatal error: unable to find a useable theme")
 				exit()
 cur_builtin = None
-#Not Sure how to explain this
-def getDefaults(folder, pluginPath=pluginPath, themePath=themePath):
-	if folder == pluginPath:
-		files = os.listdir(pluginPath)
-		defaults = []
-		for file in files:
-			try:
-				exec("global cur_builtin; from plugins." + file[:-3] + " import builtin; cur_builtin = builtin")
-				if cur_builtin == True:
-					defaults.append(file)
-			except:
-				pass
-		defaults.append("core.py")
-		defaults.append("__pycache__")
-		defaults.append(".reqs")
-		return defaults
-	elif folder == themePath:
-		themelist = os.listdir(themePath)
-		defaults = []
-		for i in range(len(themelist)):
-			cur_theme = configparser.ConfigParser()
-			cur_theme.read(themePath + "/" + themelist[i])
-			try:
-				if cur_theme["theme"]["builtin"] == "true":
-					defaults.append(themelist[i])
-			except:
-				pass
-		return defaults
 
 #Restart
 def restart():
@@ -384,19 +356,11 @@ def isPrime(num, printResult=True):
 #List Plugins
 def plugins(printval=True, hidedisabled=False):
 	plugins = os.listdir(config["paths"]["userPath"] + "/plugins/")
-	nonplugins = getDefaults(config["paths"]["userPath"] + "/plugins/")
 	j = len(plugins) - 1
 	if hidedisabled == True:
 		for i in range(j, 0, -1):
-			if plugins[i] in nonplugins or plugins[i].endswith(".disabled"):
+			if plugins[i].endswith(".disabled"):
 				plugins.remove(plugins[i])
-	else:
-		for i in range(j, 0, -1):
-			if plugins[i] in nonplugins:
-				plugins.remove(plugins[i])
-	if plugins[0] in nonplugins:
-		plugins.remove(plugins[0])
-
 	i = 0
 	if printval == True:
 		while i < len(plugins):
@@ -434,37 +398,6 @@ def shell():
 			break
 		print(os.system(cmd))
 
-# def addConfig(file, dict):
-# 	try:
-# 		with open(file, "r+") as file:
-# 			data = json.load(file)
-# 			data.update(dict)
-# 			file.seek(0)
-# 			json.dump(data, file)
-# 		return True
-# 	except ValueError:
-# 		with open(file, "r+") as file:
-# 			json.dump(dict, file)
-# 		return True
-# 	except:
-# 		return False
-#
-# def updateConfig(file, item, value):
-# 	with open(file) as f:
-# 		data = json.load(f)
-# 	try:
-# 		data[item] = value
-# 		with open(file, "r+") as f:
-# 			json.dump(data, f)
-# 		return True
-# 	except:
-# 		return False
-#
-# def readConfig(file, key):
-# 	with open(file, "r+") as f:
-# 		data = json.load(f)
-# 	return data[key]
-
 #Update wizard by tabulate
 def loadConfig():
 	items = []
@@ -474,13 +407,9 @@ def loadConfig():
 	return items
 
 def doCmdUpdate(branch="master", theme=theme):
-	nonplugins = getDefaults(config["paths"]["userPath"] + "/plugins/")
-	nonthemes = getDefaults(config["paths"]["userPath"] + "/themes/")
-
 	#Establish directories
-	plugins = str(Path(__file__).parent) + "/"
-	root = str(Path(plugins).parent) + "/"
-	themes = os.path.join(root, config["paths"]["userPath"] + "/themes/")
+	root = os.path.abspath(config["paths"]["userpath"])
+	plugins = root + "plugins/"
 	parent = str(Path(root).parent) + "/"
 	confVals = loadConfig()
 	try:
@@ -497,30 +426,25 @@ def doCmdUpdate(branch="master", theme=theme):
 		shutil.rmtree(parent + ".iithemesbackup")
 	shutil.copytree(root, parent + ".iibackup/")
 
-	#Move Plugins out of Plugins
-	os.chdir(parent)
-	tempDir = ".iipluginsbackup"
-	os.mkdir(tempDir)
-	os.chdir(plugins)
-	files = os.listdir(".")
-	for file in files:
-		if file in nonplugins:
-			continue
-		else:
+	if config["installation"]["installtype"] == "portable":
+		#Move Plugins out of Plugins
+		os.chdir(parent)
+		tempDir = ".iipluginsbackup"
+		os.mkdir(tempDir)
+		os.chdir(plugins)
+		files = os.listdir(".")
+		for file in files:
 			source = os.path.join(plugins, file)
 			dest = os.path.join(parent, tempDir)
 			shutil.move(source, dest)
 
-	#Move Themes out of Themes
-	os.chdir(parent)
-	tempThemeDir = ".iithemesbackup"
-	os.mkdir(tempThemeDir)
-	os.chdir(themes)
-	files = os.listdir(".")
-	for file in files:
-		if file in nonthemes:
-			continue
-		else:
+		#Move Themes out of Themes
+		os.chdir(parent)
+		tempThemeDir = ".iithemesbackup"
+		os.mkdir(tempThemeDir)
+		os.chdir(themes)
+		files = os.listdir(".")
+		for file in files:
 			source = os.path.join(themes, file)
 			dest = os.path.join(parent, tempThemeDir)
 			shutil.move(source, dest)
@@ -548,8 +472,10 @@ def doCmdUpdate(branch="master", theme=theme):
 		for f in os.listdir(parent + ".iibackup/"):
 			shutil.move(os.path.join(parent + ".iibackup", f), root)
 		os.rmdir(parent + ".iibackup")
-		os.rmdir(parent + ".iipluginsbackup")
-		os.rmdir(parent + ".iithemesbackup")
+		if os.path.exists(parent + ".iipluginsbackup"):
+			os.rmdir(parent + ".iipluginsbackup")
+		if os.path.exists(parent + ".iithemesbackup"):
+			os.rmdir(parent + ".iithemesbackup")
 		shutil.rmtree(parent + tempDir)
 		sys.exit("Fatal Error")
 
@@ -568,32 +494,29 @@ def doCmdUpdate(branch="master", theme=theme):
 	os.rmdir("ImaginaryInfinity-Calculator-" + branch)
 	os.remove("newcalc.zip")
 
-	#move plugins back into /plugins
-	os.chdir(parent)
-	os.chdir(tempDir)
-	files = os.listdir(".")
-	for file in files:
-		shutil.move(parent + tempDir + "/" + file, plugins)
-	os.chdir("..")
-	os.rmdir(tempDir)
-	os.chdir(root)
+	if config["installation"]["installtype"] == "portable":
+		#move plugins back into /plugins
+		os.chdir(parent)
+		os.chdir(tempDir)
+		files = os.listdir(".")
+		for file in files:
+			shutil.move(parent + tempDir + "/" + file, plugins)
+		os.chdir("..")
+		os.rmdir(tempDir)
+		os.chdir(root)
 
-	#move themes back into /themes
-	os.chdir(parent)
-	os.chdir(tempThemeDir)
-	files = os.listdir(".")
-	for file in files:
-		shutil.move(parent + tempThemeDir + "/" + file, themes)
-	os.chdir("..")
-	os.rmdir(tempThemeDir)
-	os.chdir(root)
+		#move themes back into /themes
+		os.chdir(parent)
+		os.chdir(tempThemeDir)
+		files = os.listdir(".")
+		for file in files:
+			shutil.move(parent + tempThemeDir + "/" + file, themes)
+		os.chdir("..")
+		os.rmdir(tempThemeDir)
+		os.chdir(root)
 
 	#check if all is fine
-	if os.path.isfile("main.py"):
-		pass
-	elif os.path.exists("plugins"):
-		pass
-	else:
+	if not os.path.isfile("main.py") or not os.path.exists(config["paths"]["userpath"] + "/plugins"):
 		#VERY BAD THINGS HAVE HAPPENED
 		print(theme["styles"]["error"] + "Fatal Error. Files not Found")
 		#Restore Backup
@@ -603,8 +526,7 @@ def doCmdUpdate(branch="master", theme=theme):
 		sys.exit(1)
 
 	#make launcher.sh executable
-	OS = platform.system()
-	if OS == "Linux" or OS == "Darwin" or OS == "Haiku":
+	if platform.system() == "Linux" or platform.system()S == "Darwin" or platform.system() == "Haiku":
 		os.system("chmod +x launcher.sh")
 
 	#Load old conf vals
